backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
supabase
python-dotenv
transformers
sentence-transformers
faiss-cpu
torch
accelerate
tqdm
